+++ NEWS.md
@@ -4,7 +4,9 @@
 
 ### Changes
 
+- fix ROI writing
 - `EasyFITS` replaced by `AstroFITS`.
+- `MultivariateOnlineStatistics` replaced by `OnlineSampleStatistics`.
 
 ## Version v0.3.5